Move QR code parsing to reusable function
